Created seperate statusLine class to control the status line
