+++ app/src/test/java/com/sethpthomas91/httpserver/HttpResponseWrapperTest.java
@@ -7,4 +7,5 @@ import static org.junit.Assert.*;
 
 public class HttpResponseWrapperTest {
 
+
 }
